refactor(schemas): rename and update schema imports
