+++ packages/framework/package.json
@@ -134,7 +134,9 @@
     "ts-node": "^10.9.2",
     "tsup": "^8.0.2",
     "typescript": "^5.3.3",
-    "vitest": "^1.2.1"
+    "vitest": "^1.2.1",
+    "zod": "^3.23.0",
+    "zod-to-json-schema": "^3.23.0"
   },
   "dependencies": {
     "@novu/shared": "workspace:*",
@@ -146,8 +148,6 @@
     "json-schema-faker": "^0.5.6",
     "json-schema-to-ts": "^3.0.0",
     "liquidjs": "^10.13.1",
-    "ora": "^5.4.1",
-    "zod": "^3.23.0",
-    "zod-to-json-schema": "^3.23.0"
+    "ora": "^5.4.1"
   }
 }
