+++ packages/framework/src/schemas/providers/chat/slack.schema.ts
