+++ packages/framework/src/schemas/providers/inApp/index.ts
@@ -1,6 +1,6 @@
 import { InAppProviderIdEnum } from '@novu/shared';
 import { Schema } from '../../../types';
-import { genericProviderSchemas } from '../generic';
+import { genericProviderSchemas } from '../generic.schema';
 
 export const inAppProviderSchemas = {
   [InAppProviderIdEnum.Novu]: genericProviderSchemas,
